fix: 🐛 그림을 그리는 사람이 게임에서 나갈시 다음 라운드로 넘어가도록 수정
